add integ test
