+++ internal/venice-test-common/src/integrationTest/java/com/linkedin/venice/router/TestRead.java
@@ -1,39 +1,23 @@
 package com.linkedin.venice.router;
 
-import static com.linkedin.venice.ConfigKeys.CLIENT_SYSTEM_STORE_REPOSITORY_REFRESH_INTERVAL_SECONDS;
-import static com.linkedin.venice.ConfigKeys.CLIENT_USE_SYSTEM_STORE_REPOSITORY;
-import static com.linkedin.venice.ConfigKeys.DATA_BASE_PATH;
-import static com.linkedin.venice.ConfigKeys.OFFLINE_JOB_START_TIMEOUT_MS;
-import static com.linkedin.venice.ConfigKeys.PERSISTENCE_TYPE;
 import static com.linkedin.venice.router.api.VeniceMultiKeyRoutingStrategy.HELIX_ASSISTED_ROUTING;
-import static com.linkedin.venice.router.api.VenicePathParser.TYPE_BLOB_DISCOVERY;
 import static com.linkedin.venice.router.api.VenicePathParser.TYPE_CURRENT_VERSION;
 import static com.linkedin.venice.router.api.VenicePathParser.TYPE_HEALTH_CHECK;
 import static com.linkedin.venice.router.api.VenicePathParser.TYPE_RESOURCE_STATE;
 import static com.linkedin.venice.utils.concurrent.BlockingQueueType.ARRAY_BLOCKING_QUEUE;
 import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertFalse;
-import static org.testng.Assert.assertTrue;
 import static org.testng.Assert.fail;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.linkedin.d2.balancer.D2Client;
-import com.linkedin.davinci.client.DaVinciClient;
-import com.linkedin.davinci.client.DaVinciConfig;
-import com.linkedin.davinci.client.factory.CachingDaVinciClientFactory;
 import com.linkedin.r2.transport.http.common.HttpProtocolVersion;
 import com.linkedin.venice.ConfigKeys;
-import com.linkedin.venice.D2.D2ClientUtils;
 import com.linkedin.venice.client.exceptions.VeniceClientException;
 import com.linkedin.venice.client.exceptions.VeniceClientHttpException;
 import com.linkedin.venice.client.store.AvroGenericStoreClient;
 import com.linkedin.venice.client.store.ClientConfig;
 import com.linkedin.venice.client.store.ClientFactory;
 import com.linkedin.venice.client.store.ComputeGenericRecord;
-import com.linkedin.venice.common.VeniceSystemStoreType;
-import com.linkedin.venice.common.VeniceSystemStoreUtils;
-import com.linkedin.venice.controllerapi.BlobDiscoveryResponse;
-import com.linkedin.venice.controllerapi.ControllerClient;
 import com.linkedin.venice.controllerapi.ControllerResponse;
 import com.linkedin.venice.controllerapi.CurrentVersionResponse;
 import com.linkedin.venice.controllerapi.D2ServiceDiscoveryResponse;
@@ -43,15 +27,10 @@ import com.linkedin.venice.controllerapi.VersionCreationResponse;
 import com.linkedin.venice.exceptions.VeniceException;
 import com.linkedin.venice.helix.HelixReadOnlySchemaRepository;
 import com.linkedin.venice.integration.utils.D2TestUtils;
-import com.linkedin.venice.integration.utils.PubSubBrokerWrapper;
 import com.linkedin.venice.integration.utils.ServiceFactory;
 import com.linkedin.venice.integration.utils.VeniceClusterWrapper;
-import com.linkedin.venice.integration.utils.VeniceControllerWrapper;
-import com.linkedin.venice.integration.utils.VeniceMultiClusterWrapper;
 import com.linkedin.venice.integration.utils.VeniceRouterWrapper;
 import com.linkedin.venice.integration.utils.VeniceServerWrapper;
-import com.linkedin.venice.integration.utils.VeniceTwoLayerMultiRegionMultiClusterWrapper;
-import com.linkedin.venice.meta.PersistenceType;
 import com.linkedin.venice.meta.Version;
 import com.linkedin.venice.pubsub.PubSubProducerAdapterFactory;
 import com.linkedin.venice.router.api.VenicePathParser;
@@ -63,12 +42,10 @@ import com.linkedin.venice.stats.StatsErrorCode;
 import com.linkedin.venice.tehuti.MetricsUtils;
 import com.linkedin.venice.utils.IntegrationTestPushUtils;
 import com.linkedin.venice.utils.ObjectMapperFactory;
-import com.linkedin.venice.utils.PropertyBuilder;
 import com.linkedin.venice.utils.SslUtils;
 import com.linkedin.venice.utils.TestUtils;
 import com.linkedin.venice.utils.Time;
 import com.linkedin.venice.utils.Utils;
-import com.linkedin.venice.utils.VeniceProperties;
 import com.linkedin.venice.writer.VeniceWriter;
 import com.linkedin.venice.writer.VeniceWriterOptions;
 import io.netty.handler.codec.http.HttpResponseStatus;
@@ -76,20 +53,13 @@ import io.tehuti.Metric;
 import io.tehuti.metrics.MetricsRepository;
 import java.io.IOException;
 import java.io.InputStream;
-import java.lang.reflect.Method;
-import java.util.AbstractMap;
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Properties;
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 import org.apache.avro.Schema;
 import org.apache.avro.generic.GenericData;
 import org.apache.avro.generic.GenericRecord;
@@ -105,7 +75,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 
@@ -118,19 +88,12 @@ public abstract class TestRead {
   private String storeVersionName;
   private int valueSchemaId;
   private String storeName;
-  private String[] clusterNames;
-  private String parentControllerURLs;
-  private VeniceMultiClusterWrapper multiClusterVenice;
   private final String readDisabledStoreName = Utils.getUniqueString("read_disabled_store");
 
   private String routerAddr;
   private VeniceKafkaSerializer keySerializer;
   private VeniceKafkaSerializer valueSerializer;
   private VeniceWriter<Object, Object, Object> veniceWriter;
-  private VeniceTwoLayerMultiRegionMultiClusterWrapper multiRegionMultiClusterWrapper;
-
-  private static final String INT_KEY_SCHEMA = "\"int\"";
-  private static final String INT_VALUE_SCHEMA = "\"int\"";
   private static final String KEY_SCHEMA_STR = "\"string\"";
   private static final String VALUE_FIELD_NAME = "int_field";
   private static final String UNUSED_FIELD_NAME = "unused_field";
@@ -158,56 +121,11 @@ public abstract class TestRead {
     return veniceCluster;
   }
 
-  @BeforeMethod
-  public void setUp(Method method) throws VeniceClientException, ExecutionException, InterruptedException {
+  @BeforeClass(alwaysRun = true)
+  public void setUp() throws VeniceClientException, ExecutionException, InterruptedException {
     if (!isTestEnabled()) {
       return;
     }
-    if ("testBlobDiscovery".equals(method.getName())) {
-      setUpForBlobTransfer();
-    } else {
-      setUpTestRead();
-    }
-  }
-
-  public void setUpForBlobTransfer() {
-    Utils.thisIsLocalhost();
-    Properties parentControllerProps = new Properties();
-    parentControllerProps.put(OFFLINE_JOB_START_TIMEOUT_MS, "180000");
-    multiRegionMultiClusterWrapper = ServiceFactory.getVeniceTwoLayerMultiRegionMultiClusterWrapper(
-        1,
-        2,
-        1,
-        1,
-        3,
-        1,
-        3,
-        Optional.of(parentControllerProps),
-        Optional.empty(),
-        Optional.empty(),
-        false);
-    multiClusterVenice = multiRegionMultiClusterWrapper.getChildRegions().get(0);
-    clusterNames = multiClusterVenice.getClusterNames();
-    parentControllerURLs = multiRegionMultiClusterWrapper.getParentControllers()
-        .stream()
-        .map(VeniceControllerWrapper::getControllerUrl)
-        .collect(Collectors.joining(","));
-
-    for (String cluster: clusterNames) {
-      try (ControllerClient controllerClient =
-          new ControllerClient(cluster, multiClusterVenice.getControllerConnectString())) {
-        // Verify the participant store is up and running in child region
-        String participantStoreName = VeniceSystemStoreUtils.getParticipantStoreNameForCluster(cluster);
-        TestUtils.waitForNonDeterministicPushCompletion(
-            Version.composeKafkaTopic(participantStoreName, 1),
-            controllerClient,
-            5,
-            TimeUnit.MINUTES);
-      }
-    }
-  }
-
-  public void setUpTestRead() throws VeniceClientException, ExecutionException, InterruptedException {
     /**
      * The following config is used to detect Netty resource leaking.
      * If memory leak happens, you will see the following log message:
@@ -225,7 +143,7 @@ public abstract class TestRead {
     // To trigger long-tail retry
     extraProperties.put(ConfigKeys.ROUTER_LONG_TAIL_RETRY_FOR_SINGLE_GET_THRESHOLD_MS, 1);
     extraProperties.put(ConfigKeys.ROUTER_MAX_KEY_COUNT_IN_MULTIGET_REQ, MAX_KEY_LIMIT); // 20 keys at most in a
-                                                                                         // batch-get request
+    // batch-get request
     extraProperties.put(ConfigKeys.ROUTER_LONG_TAIL_RETRY_FOR_BATCH_GET_THRESHOLD_MS, "1-:1");
     extraProperties.put(ConfigKeys.ROUTER_SMART_LONG_TAIL_RETRY_ENABLED, false);
     extraProperties.put(ConfigKeys.ROUTER_STORAGE_NODE_CLIENT_TYPE, getStorageNodeClientType());
@@ -292,7 +210,6 @@ public abstract class TestRead {
         throw new VeniceException(
             "Failed to create a store: " + readDisabledStoreName + " with error: " + newStoreResponse.getError());
       }
-
       VersionCreationResponse versionCreationResponse = cc.emptyPush(readDisabledStoreName, "test_push", 10000);
       if (versionCreationResponse.isError()) {
         throw new VeniceException(
@@ -343,7 +260,6 @@ public abstract class TestRead {
     }
     Utils.closeQuietlyWithErrorLogged(veniceCluster);
     Utils.closeQuietlyWithErrorLogged(veniceWriter);
-    Utils.closeQuietlyWithErrorLogged(multiRegionMultiClusterWrapper);
     if (d2Client != null) {
       d2Client.shutdown(null);
     }
@@ -412,26 +328,26 @@ public abstract class TestRead {
 
         double maxInflightRequestCountAfterQueries =
             getAggregateRouterMetricValue(".total--in_flight_request_count.Max");
-        assertTrue(maxInflightRequestCountAfterQueries > 0.0, "There should be in-flight requests now!");
+        Assert.assertTrue(maxInflightRequestCountAfterQueries > 0.0, "There should be in-flight requests now!");
 
         // Check retry requests
-        assertTrue(
+        Assert.assertTrue(
             getAggregateRouterMetricValue(".total--retry_count.LambdaStat") > 0,
             "After " + rounds + " reads, there should be some single-get retry requests");
-        assertTrue(
+        Assert.assertTrue(
             getAggregateRouterMetricValue(".total--retry_delay.Avg") > 0,
             "After " + rounds + " reads, there should be some single-get retry requests");
-        assertTrue(
+        Assert.assertTrue(
             getAggregateRouterMetricValue(".total--multiget_streaming_retry_count.LambdaStat") > 0,
             "After " + rounds + " reads, there should be some batch-get retry requests");
 
         // Check Router connection pool metrics
         if (getStorageNodeClientType() == StorageNodeClientType.APACHE_HTTP_ASYNC_CLIENT) {
           // TODO: add connection pool stats for netty client
-          assertTrue(
+          Assert.assertTrue(
               getAggregateRouterMetricValue(".connection_pool--total_max_connection_count.LambdaStat") > 0,
               "Max connection count must be positive");
-          assertTrue(
+          Assert.assertTrue(
               getMaxRouterMetricValue(".connection_pool--connection_lease_request_latency.Max") > 0,
               "Connection lease max latency should be positive");
           assertEquals(
@@ -442,11 +358,11 @@ public abstract class TestRead {
               getAggregateRouterMetricValue(".connection_pool--total_pending_connection_request_count.LambdaStat"),
               0.0d,
               "Pending connection request count should be 0 since test queries are finished");
-          assertTrue(
+          Assert.assertTrue(
               getAggregateRouterMetricValue(".connection_pool--total_idle_connection_count.LambdaStat") > 0,
               "There should be some idle connections since test queries are finished");
 
-          assertTrue(
+          Assert.assertTrue(
               getAggregateRouterMetricValue(".localhost--max_connection_count.Gauge") > 0,
               "Max connection count must be positive");
           assertEquals(
@@ -457,17 +373,17 @@ public abstract class TestRead {
               getAggregateRouterMetricValue(".localhost--pending_connection_request_count.Gauge"),
               0.0d,
               "Pending connection request count should be 0 since test queries are finished");
-          assertTrue(
+          Assert.assertTrue(
               getAggregateRouterMetricValue(".localhost--idle_connection_count.Gauge") > 0,
               "There should be some idle connections since test queries are finished");
         }
 
-        assertTrue(getAggregateRouterMetricValue(".localhost--response_waiting_time.50thPercentile") > 0);
-        assertTrue(
+        Assert.assertTrue(getAggregateRouterMetricValue(".localhost--response_waiting_time.50thPercentile") > 0);
+        Assert.assertTrue(
             getAggregateRouterMetricValue(".localhost--multiget_streaming_response_waiting_time.50thPercentile") > 0);
 
-        assertTrue(getAggregateRouterMetricValue(".localhost--request.Count") > 0);
-        assertTrue(getAggregateRouterMetricValue(".localhost--multiget_streaming_request.Count") > 0);
+        Assert.assertTrue(getAggregateRouterMetricValue(".localhost--request.Count") > 0);
+        Assert.assertTrue(getAggregateRouterMetricValue(".localhost--multiget_streaming_request.Count") > 0);
 
         // Each round:
         // 1. We do MAX_KEY_LIMIT * 2 because we do a batch get and a batch compute
@@ -484,7 +400,8 @@ public abstract class TestRead {
           Assert.assertEquals(getMaxServerMetricValue(".total--multiget_request_part_count.Max"), 1.0);
           if (readComputeEnabled) {
             Assert.assertEquals(getMaxServerMetricValue(".total--compute_request_part_count.Max"), 1.0);
-            assertTrue(getMaxServerMetricValue(".total--compute_storage_engine_read_compute_efficiency.Max") > 1.0);
+            Assert.assertTrue(
+                getMaxServerMetricValue(".total--compute_storage_engine_read_compute_efficiency.Max") > 1.0);
             Assert.assertEquals(getAggregateRouterMetricValue(".total--compute_multiget_fallback.Total"), 0.0);
             Assert.assertEquals(
                 clientMetrics.getMetric("." + storeName + "--compute_streaming_multiget_fallback.OccurrenceRate")
@@ -494,16 +411,16 @@ public abstract class TestRead {
             Assert.assertEquals(
                 getAggregateRouterMetricValue(".total--compute_multiget_fallback.Total"),
                 (double) MAX_KEY_LIMIT);
-            assertTrue(
+            Assert.assertTrue(
                 clientMetrics.getMetric("." + storeName + "--compute_streaming_multiget_fallback.OccurrenceRate")
                     .value() > 0);
           }
         }
 
         // Verify storage node metrics
-        assertTrue(getMaxServerMetricValue(".total--records_consumed.Rate") > 0.0);
-        assertTrue(getMaxServerMetricValue(".total--multiget_request_size_in_bytes.Max") > 0.0);
-        assertTrue(getMaxServerMetricValue(".total--compute_request_size_in_bytes.Max") > 0.0);
+        Assert.assertTrue(getMaxServerMetricValue(".total--records_consumed.Rate") > 0.0);
+        Assert.assertTrue(getMaxServerMetricValue(".total--multiget_request_size_in_bytes.Max") > 0.0);
+        Assert.assertTrue(getMaxServerMetricValue(".total--compute_request_size_in_bytes.Max") > 0.0);
 
         for (VeniceServerWrapper veniceServerWrapper: veniceCluster.getVeniceServers()) {
           Map<String, ? extends Metric> metrics = veniceServerWrapper.getMetricsRepository().metrics();
@@ -518,7 +435,7 @@ public abstract class TestRead {
                   value,
                   (double) StatsErrorCode.NULL_STORAGE_ENGINE_STATS.code,
                   "Got NULL_STORAGE_ENGINE_STATS!");
-              assertTrue(value > 0.0, "Disk usage for current version should be positive. Got: " + value);
+              Assert.assertTrue(value > 0.0, "Disk usage for current version should be positive. Got: " + value);
             }
           });
         }
@@ -571,7 +488,7 @@ public abstract class TestRead {
           storeClient.get(KEY_PREFIX + i).get();
         } catch (ExecutionException e) {
           Throwable cause = e.getCause();
-          assertTrue(cause instanceof VeniceClientHttpException);
+          Assert.assertTrue(cause instanceof VeniceClientHttpException);
           Assert.assertEquals(
               ((VeniceClientHttpException) cause).getHttpStatus(),
               HttpResponseStatus.TOO_MANY_REQUESTS.code());
@@ -579,7 +496,7 @@ public abstract class TestRead {
         }
       }
       long runTimeMs = System.currentTimeMillis() - startTime;
-      assertTrue(
+      Assert.assertTrue(
           quotaExceptionsCount > 0,
           "There were no quota exceptions at all for single gets! " + "(Test too slow? " + runTimeMs + " ms for "
               + numberOfRequests + " requests)");
@@ -626,10 +543,10 @@ public abstract class TestRead {
           storeClient.batchGet(keySet).get();
         } catch (ExecutionException e) {
           Throwable cause = e.getCause();
-          assertTrue(
+          Assert.assertTrue(
               cause instanceof VeniceClientHttpException,
               "Wanted " + VeniceClientHttpException.class.getSimpleName() + " but instead got: " + cause);
-          assertTrue(
+          Assert.assertTrue(
               cause.getMessage().contains("Quota exceeded"),
               "Did not get the expected exception message: " + cause.getMessage());
           quotaExceptionsCountForBatchGet++;
@@ -650,7 +567,7 @@ public abstract class TestRead {
           runTimeForBatchGetMs,
           queriesSent,
           quotaExceptionsCountForBatchGet);
-      assertTrue(
+      Assert.assertTrue(
           quotaExceptionsCountForBatchGet > 0,
           "There were no quota exceptions at all for batch gets! " + "(Test too slow? " + runTimeForBatchGetMs
               + " ms for " + numberOfRequests + " requests)");
@@ -707,7 +624,7 @@ public abstract class TestRead {
       ObjectMapper mapper = ObjectMapperFactory.getInstance();
       D2ServiceDiscoveryResponse d2ServiceDiscoveryResponse =
           mapper.readValue(responseBody.getBytes(), D2ServiceDiscoveryResponse.class);
-      assertFalse(d2ServiceDiscoveryResponse.isError());
+      Assert.assertFalse(d2ServiceDiscoveryResponse.isError());
       Assert.assertEquals(d2ServiceDiscoveryResponse.getCluster(), veniceCluster.getClusterName());
       Assert.assertEquals(
           d2ServiceDiscoveryResponse.getD2Service(),
@@ -835,158 +752,5 @@ public abstract class TestRead {
         .toString();
     assertEquals(metricsRepository.metrics().get(requestUsageMetric).value(), 1.0d);
     assertEquals(metricsRepository.metrics().get(badRequestMetric).value(), 1.0d);
-  }
-
-  @Test(timeOut = 10 * Time.MINUTES_PER_HOUR)
-  public void testBlobDiscovery() throws Exception {
-    if (!isTestEnabled()) {
-      return;
-    }
-
-    assertTrue(clusterNames.length > 1, "Insufficient clusters for this test to be meaningful");
-    int initialKeyCount = 10;
-    List<String> stores = new ArrayList<>();
-    // Create a new store in each cluster and setup their corresponding meta system store.
-    List<PubSubBrokerWrapper> pubSubBrokerWrappers = multiClusterVenice.getClusters()
-        .values()
-        .stream()
-        .map(VeniceClusterWrapper::getPubSubBrokerWrapper)
-        .collect(Collectors.toList());
-    Map<String, String> additionalPubSubProperties =
-        PubSubBrokerWrapper.getBrokerDetailsForClients(pubSubBrokerWrappers);
-
-    for (int index = 0; index < clusterNames.length; index++) {
-      final int value = index;
-      String cluster = clusterNames[index];
-      // Create the venice stores and materialize the corresponding meta system store for each store.
-      try (ControllerClient parentControllerClient = new ControllerClient(cluster, parentControllerURLs)) {
-        String storeName = Utils.getUniqueString("test-store");
-        stores.add(storeName);
-
-        assertFalse(
-            parentControllerClient.createNewStore(storeName, "venice-test", INT_KEY_SCHEMA, INT_VALUE_SCHEMA)
-                .isError());
-
-        PubSubProducerAdapterFactory pubSubProducerAdapterFactory = multiClusterVenice.getClusters()
-            .get(cluster)
-            .getPubSubBrokerWrapper()
-            .getPubSubClientsFactory()
-            .getProducerAdapterFactory();
-        VersionCreationResponse response = TestUtils.createVersionWithBatchData(
-            parentControllerClient,
-            storeName,
-            INT_KEY_SCHEMA,
-            INT_VALUE_SCHEMA,
-            IntStream.range(0, initialKeyCount).mapToObj(i -> new AbstractMap.SimpleEntry<>(i, value)),
-            pubSubProducerAdapterFactory,
-            additionalPubSubProperties);
-        // Verify the data can be ingested by classical Venice before proceeding.
-        TestUtils.waitForNonDeterministicPushCompletion(
-            response.getKafkaTopic(),
-            parentControllerClient,
-            30,
-            TimeUnit.SECONDS);
-        makeSureSystemStoresAreOnline(parentControllerClient, storeName);
-        multiClusterVenice.getClusters().get(cluster).refreshAllRouterMetaData();
-      }
-    }
-
-    VeniceProperties backendConfig =
-        new PropertyBuilder().put(DATA_BASE_PATH, Utils.getTempDataDirectory().getAbsolutePath())
-            .put(PERSISTENCE_TYPE, PersistenceType.ROCKS_DB)
-            .put(CLIENT_USE_SYSTEM_STORE_REPOSITORY, true)
-            .put(CLIENT_SYSTEM_STORE_REPOSITORY_REFRESH_INTERVAL_SECONDS, 1)
-            .build();
-    DaVinciConfig daVinciConfig = new DaVinciConfig();
-    D2Client daVinciD2 = D2TestUtils.getAndStartD2Client(multiClusterVenice.getZkServerWrapper().getAddress());
-
-    try (CachingDaVinciClientFactory factory = new CachingDaVinciClientFactory(
-        daVinciD2,
-        VeniceRouterWrapper.CLUSTER_DISCOVERY_D2_SERVICE_NAME,
-        new MetricsRepository(),
-        backendConfig)) {
-      List<DaVinciClient<Integer, Object>> clients = new ArrayList<>();
-      for (int i = 0; i < stores.size(); i++) {
-        String store = stores.get(i);
-        DaVinciClient<Integer, Object> client = factory.getAndStartGenericAvroClient(store, daVinciConfig);
-        client.subscribeAll().get();
-        for (int k = 0; k < initialKeyCount; k++) {
-          assertEquals(client.get(k).get(), i);
-        }
-        clients.add(client);
-      }
-      // Verify new push works
-      final int newValue = 1000;
-      try (ControllerClient parentControllerClient = new ControllerClient(clusterNames[0], parentControllerURLs)) {
-        PubSubProducerAdapterFactory pubSubProducerAdapterFactory = multiClusterVenice.getClusters()
-            .get(clusterNames[0])
-            .getPubSubBrokerWrapper()
-            .getPubSubClientsFactory()
-            .getProducerAdapterFactory();
-        VersionCreationResponse versionCreationResponse = TestUtils.createVersionWithBatchData(
-            parentControllerClient,
-            stores.get(0),
-            INT_KEY_SCHEMA,
-            INT_VALUE_SCHEMA,
-            IntStream.range(0, initialKeyCount).mapToObj(i -> new AbstractMap.SimpleEntry<>(i, newValue)),
-            pubSubProducerAdapterFactory,
-            additionalPubSubProperties);
-        TestUtils.waitForNonDeterministicPushCompletion(
-            versionCreationResponse.getKafkaTopic(),
-            parentControllerClient,
-            60,
-            TimeUnit.SECONDS);
-      }
-      TestUtils.waitForNonDeterministicAssertion(2, TimeUnit.MINUTES, true, () -> {
-        for (int k = 0; k < initialKeyCount; k++) {
-          assertEquals(clients.get(0).get(k).get(), newValue);
-        }
-      });
-
-      String aNewStoreName = stores.get(0);
-      VeniceClusterWrapper vcw = multiClusterVenice.getClusters().get(clusterNames[0]);
-
-      TestUtils.waitForNonDeterministicAssertion(2, TimeUnit.MINUTES, true, () -> {
-        vcw.updateStore(aNewStoreName, new UpdateStoreQueryParams().setBlobTransferEnabled(true));
-      });
-
-      String routerURL = vcw.getRandomRouterURL();
-      try (CloseableHttpAsyncClient client = HttpAsyncClients.custom()
-          .setDefaultRequestConfig(RequestConfig.custom().setSocketTimeout(2000).build())
-          .build()) {
-        client.start();
-
-        String uri = routerURL + "/" + TYPE_BLOB_DISCOVERY + "?store_name=" + aNewStoreName
-            + "&store_version=1&store_partition=1";
-        HttpGet routerRequest = new HttpGet(uri);
-        HttpResponse response = client.execute(routerRequest, null).get();
-        String responseBody;
-        try (InputStream bodyStream = response.getEntity().getContent()) {
-          responseBody = IOUtils.toString(bodyStream);
-        }
-        Assert.assertEquals(
-            response.getStatusLine().getStatusCode(),
-            HttpStatus.SC_OK,
-            "Failed to get resource state for " + storeVersionName + ". Response: " + responseBody);
-        ObjectMapper mapper = ObjectMapperFactory.getInstance();
-        BlobDiscoveryResponse blobDiscoveryResponse =
-            mapper.readValue(responseBody.getBytes(), BlobDiscoveryResponse.class);
-        Assert.assertEquals(blobDiscoveryResponse.getLiveNodeHostNames().size(), 0);
-        LOGGER.info(responseBody);
-      } catch (Exception e) {
-        fail("Unexpected exception", e);
-      }
-    } finally {
-      D2ClientUtils.shutdownClient(daVinciD2);
-    }
-  }
-
-  private void makeSureSystemStoresAreOnline(ControllerClient controllerClient, String storeName) {
-    String metaSystemStoreTopic =
-        Version.composeKafkaTopic(VeniceSystemStoreType.META_STORE.getSystemStoreName(storeName), 1);
-    TestUtils.waitForNonDeterministicPushCompletion(metaSystemStoreTopic, controllerClient, 30, TimeUnit.SECONDS);
-    String daVinciPushStatusStore =
-        Version.composeKafkaTopic(VeniceSystemStoreType.DAVINCI_PUSH_STATUS_STORE.getSystemStoreName(storeName), 1);
-    TestUtils.waitForNonDeterministicPushCompletion(daVinciPushStatusStore, controllerClient, 30, TimeUnit.SECONDS);
   }
 }
